+++ lib/components/onboarding/dot_indicator.dart
