+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:greenit_version1/screens/main_navigation.dart';
 import 'package:greenit_version1/screens/onboarding/onboarding_screen.dart';
 import 'package:greenit_version1/size_config.dart';
 import 'package:greenit_version1/theme.dart';
