--- lib/screens/home/components/google_map_api.dart
@@ -1,23 +1,0 @@
-// import 'package:flutter/material.dart';
-
-// class GoogleMapApi extends StatefulWidget {
-//   const GoogleMapApi({super.key});
-
-//   @override
-//   State<GoogleMapApi> createState() => _GoogleMapApiState();
-// }
-
-// class _GoogleMapApiState extends State<GoogleMapApi> {
-//   final Completer<GoogleMapController> _controller =
-//       Completer<GoogleMapController>();
-
-//   static const CameraPosition _kGooglePlex = CameraPosition(
-//     target: LatLng(37.42796133580664, -122.085749655962),
-//     zoom: 14.4746,
-//   );
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return const Placeholder();
-//   }
-// }
