+++ lib/screens/otpVerification/components/body.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:form_field_validator/form_field_validator.dart';
-import 'package:greenit_version1/components/buttons/primary_button.dart';
-import 'package:greenit_version1/components/welcome_text.dart';
+import 'package:greenit_version1/components/forms/welcome_text.dart';
 import 'package:greenit_version1/constants.dart';
-import 'package:greenit_version1/screens/successLogin/success_login_screen.dart';
+import 'package:greenit_version1/screens/otpVerification/components/otp_verification_form.dart';
 import 'package:greenit_version1/size_config.dart';
 
 class Body extends StatelessWidget {
@@ -29,163 +27,6 @@ class Body extends StatelessWidget {
           ),
         ),
       ),
-    );
-  }
-}
-
-class OtpVerificationForm extends StatefulWidget {
-  const OtpVerificationForm({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  State<OtpVerificationForm> createState() => _OtpVerificationFormState();
-}
-
-class _OtpVerificationFormState extends State<OtpVerificationForm> {
-  final _formKey = GlobalKey<FormState>();
-
-  FocusNode? pin2FocusNode;
-  FocusNode? pin3FocusNode;
-  FocusNode? pin4FocusNode;
-
-  @override
-  void initState() {
-    super.initState();
-    pin2FocusNode = FocusNode();
-    pin3FocusNode = FocusNode();
-    pin4FocusNode = FocusNode();
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    pin2FocusNode!.dispose();
-    pin3FocusNode!.dispose();
-    pin4FocusNode!.dispose();
-  }
-
-  void nextField(String value, FocusNode? focusNode) {
-    if (value.length == 1) {
-      focusNode!.requestFocus();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-      key: _formKey,
-      child: Column(
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              SizedBox(
-                width: getProportionateScreenWidth(60),
-                child: TextFormField(
-                  autofocus: true,
-                  maxLength: 1,
-                  validator: RequiredValidator(errorText: ''),
-                  style: const TextStyle(fontSize: 24),
-                  keyboardType: TextInputType.number,
-                  textAlign: TextAlign.center,
-                  decoration: otpInputDecoration,
-                  onChanged: (value) {
-                    nextField(value, pin2FocusNode);
-                  },
-                ),
-              ),
-              SizedBox(
-                width: getProportionateScreenWidth(60),
-                child: TextFormField(
-                  focusNode: pin2FocusNode,
-                  validator: RequiredValidator(errorText: ''),
-                  maxLength: 1,
-                  style: const TextStyle(fontSize: 24),
-                  keyboardType: TextInputType.number,
-                  textAlign: TextAlign.center,
-                  decoration: otpInputDecoration,
-                  onChanged: (value) => nextField(value, pin3FocusNode),
-                ),
-              ),
-              SizedBox(
-                width: getProportionateScreenWidth(60),
-                child: TextFormField(
-                  focusNode: pin3FocusNode,
-                  validator: RequiredValidator(errorText: ''),
-                  maxLength: 1,
-                  style: const TextStyle(fontSize: 24),
-                  keyboardType: TextInputType.number,
-                  textAlign: TextAlign.center,
-                  decoration: otpInputDecoration,
-                  onChanged: (value) => nextField(value, pin4FocusNode),
-                ),
-              ),
-              SizedBox(
-                width: getProportionateScreenWidth(60),
-                child: TextFormField(
-                  focusNode: pin4FocusNode,
-                  validator: RequiredValidator(errorText: ''),
-                  maxLength: 1,
-                  style: const TextStyle(fontSize: 24),
-                  keyboardType: TextInputType.number,
-                  textAlign: TextAlign.center,
-                  decoration: otpInputDecoration,
-                  onChanged: (value) {
-                    if (value.length == 1) {
-                      pin4FocusNode!.unfocus();
-                    }
-                  },
-                ),
-              ),
-            ],
-          ),
-          VerticalSpacing(of: SizeConfig.screenHeight * 0.15),
-          PrimaryButton(
-            text: "Continue",
-            press: () {
-              if (_formKey.currentState!.validate()) {
-                _formKey.currentState!.save();
-
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) => const SuccessLoginScreen(),
-                  ),
-                );
-              }
-            },
-          ),
-          VerticalSpacing(of: SizeConfig.screenHeight * 0.1),
-          const ResendCodeText()
-        ],
-      ),
-    );
-  }
-}
-
-class ResendCodeText extends StatelessWidget {
-  const ResendCodeText({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Text(
-          "Didn't receive code? ",
-          style: kBodyTextStyle,
-        ),
-        GestureDetector(
-          onTap: () {},
-          child: Text(
-            'Resend Again.',
-            style: kBodyTextStyle.copyWith(color: kPrimaryActiveColor),
-          ),
-        ),
-      ],
     );
   }
 }
